Split resolwe.flow.testcases module into multiple modules
This will avoid resolwe.flow.testcases module becoming too long (i.e.
more than 1000 lines).
