+++ resolwe/test/__init__.py
@@ -12,13 +12,16 @@ Resolwe Test Framework
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from resolwe.test.testcases import (
-    ElasticSearchTestCase, ProcessTestCase, ResolweAPITestCase, TestCase, TransactionElasticSearchTestCase,
-    TransactionTestCase,
-)
+from resolwe.test.testcases import TestCase, TransactionTestCase
+from resolwe.test.testcases.api import ResolweAPITestCase
+from resolwe.test.testcases.elastic import ElasticSearchTestCase, TransactionElasticSearchTestCase
+from resolwe.test.testcases.process import ProcessTestCase
 from resolwe.test.utils import check_docker, check_installed, with_docker_executor
 
 __all__ = (
-    'TestCase', 'TransactionTestCase', 'ProcessTestCase', 'ResolweAPITestCase', 'ElasticSearchTestCase',
-    'TransactionElasticSearchTestCase', 'check_installed', 'check_docker', 'with_docker_executor',
+    'TestCase', 'TransactionTestCase',
+    'ResolweAPITestCase',
+    'ElasticSearchTestCase', 'TransactionElasticSearchTestCase',
+    'ProcessTestCase',
+    'check_docker', 'check_installed', 'with_docker_executor',
 )
