+++ resolwe/test/testcases/process.py
@@ -1,31 +1,12 @@
 """.. Ignore pydocstyle D400.
 
-==================
-Resolwe Test Cases
-==================
-
-.. autoclass:: resolwe.test.TransactionTestCase
-    :members:
-
-.. autoclass:: resolwe.test.TestCase
-    :members:
-
 .. autoclass:: resolwe.test.ProcessTestCase
     :members:
 
-.. autoclass:: resolwe.test.ResolweAPITestCase
-    :members:
-
-.. autoclass:: resolwe.test.TransactionElasticSearchTestCase
-    :members:
-
-.. autoclass:: resolwe.test.ElasticSearchTestCase
-    :members:
-
 """
+
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import copy
 import gzip
 import hashlib
 import io
@@ -35,165 +16,23 @@ import shutil
 import uuid
 import zipfile
 
-import mock
-import six
 from six.moves import filterfalse
 
 from django.conf import settings
-from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Group
 from django.core import management
-from django.core.urlresolvers import reverse
-from django.test import TestCase as DjangoTestCase
-from django.test import TransactionTestCase as DjangoTransactionTestCase
 from django.test import override_settings
 from django.utils.crypto import get_random_string
 from django.utils.text import slugify
 
-from rest_framework.test import APIRequestFactory, APITestCase, force_authenticate
-
 from resolwe.flow.managers import manager
 from resolwe.flow.models import Collection, Data, DescriptorSchema, Process, Storage
 from resolwe.flow.utils import dict_dot, iterate_fields, iterate_schema
+from resolwe.test import TestCase
 
-if six.PY2:
-    # Monkey-patch shutil package with which function (available in Python 3.3+)
-    import shutilwhich  # pylint: disable=import-error,unused-import
-
-__all__ = (
-    'TestCase', 'TransactionTestCase', 'ProcessTestCase', 'ResolweAPITestCase',
-    'ElasticSearchTestCase', 'TransactionElasticSearchTestCase',
-)
+from .setting_overrides import FLOW_DOCKER_MAPPINGS, FLOW_EXECUTOR_SETTINGS
 
 
 SCHEMAS_FIXTURE_CACHE = None
-
-# override all FLOW_EXECUTOR settings that are specified in FLOW_EXECUTOR['TEST']
-FLOW_EXECUTOR_SETTINGS = copy.copy(getattr(settings, 'FLOW_EXECUTOR', {}))
-TEST_SETTINGS_OVERRIDES = getattr(settings, 'FLOW_EXECUTOR', {}).get('TEST', {})
-FLOW_EXECUTOR_SETTINGS.update(TEST_SETTINGS_OVERRIDES)
-
-# update FLOW_DOCKER_MAPPINGS setting if necessary
-FLOW_DOCKER_MAPPINGS = copy.copy(getattr(settings, 'FLOW_DOCKER_MAPPINGS', {}))
-for map_ in FLOW_DOCKER_MAPPINGS:
-    for map_entry in ['src', 'dest']:
-        for setting in ['DATA_DIR', 'UPLOAD_DIR']:
-            if settings.FLOW_EXECUTOR[setting] in map_[map_entry]:
-                map_[map_entry] = map_[map_entry].replace(
-                    settings.FLOW_EXECUTOR[setting], FLOW_EXECUTOR_SETTINGS[setting], 1)
-
-
-class TransactionTestCase(DjangoTransactionTestCase):
-    """Base class for writing Resolwe tests not enclosed in a transaction.
-
-    It is based on Django's :class:`~django.test.TransactionTestCase`.
-    Use it if you need to access the test's database from another
-    thread/process.
-
-    """
-
-    def _test_data_dir(self, path):
-        """Return test data directory path.
-
-        Increase counter in the path name by 1.
-
-        """
-        while True:
-            try:
-                counter = 1
-                for name in os.listdir(path):
-                    if os.path.isdir(os.path.join(path, name)) and name.startswith('test'):
-                        try:
-                            current = int(name.split('_')[-1])
-                            if current >= counter:
-                                counter = current + 1
-                        except ValueError:
-                            pass
-
-                test_data_dir = os.path.join(path, 'test_{}'.format(counter))
-                os.makedirs(test_data_dir)
-                break
-            except OSError:
-                # Try again if a folder with the same name was created
-                # by another test on another thread
-                continue
-
-        return test_data_dir
-
-    def setUp(self):
-        """Initialize test data."""
-        super(TransactionTestCase, self).setUp()
-
-        # Override flow executor settings
-        flow_executor_settings = copy.copy(getattr(settings, 'FLOW_EXECUTOR', {}))
-
-        # Override data directory settings
-        data_dir = self._test_data_dir(FLOW_EXECUTOR_SETTINGS['DATA_DIR'])
-        flow_executor_settings['DATA_DIR'] = data_dir
-
-        # Override container name prefix setting
-        flow_executor_settings['CONTAINER_NAME_PREFIX'] = '{}_{}_{}'.format(
-            getattr(settings, 'FLOW_EXECUTOR', {}).get('CONTAINER_NAME_PREFIX', 'resolwe'),
-            # NOTE: This is necessary to avoid container name clashes when tests are run from
-            # different Resolwe code bases on the same system (e.g. on a CI server).
-            get_random_string(length=6),
-            os.path.basename(data_dir)
-        )
-
-        # Override Docker data directory mappings
-        flow_docker_mappings = copy.copy(getattr(settings, 'FLOW_DOCKER_MAPPINGS', []))
-        for mapping in flow_docker_mappings:
-            if mapping['dest'] == '/home/biolinux/data':
-                mapping['src'] = os.path.join(data_dir, '{data_id}')
-            elif mapping['dest'] == '/data_all':
-                mapping['src'] = data_dir
-
-        self.settings = override_settings(FLOW_EXECUTOR=flow_executor_settings,
-                                          FLOW_DOCKER_MAPPINGS=flow_docker_mappings)
-        self.settings.enable()
-
-        self._keep_data = False
-
-        user_model = get_user_model()
-        self.admin = user_model.objects.create_superuser(username='admin', email='admin@test.com', password='admin')
-        self.contributor = user_model.objects.create_user(username='contributor')
-        self.user = user_model.objects.create_user(username='normal_user')
-
-        self.group = Group.objects.create(name='Users')
-        self.group.user_set.add(self.user)
-
-    def tearDown(self):
-        """Clean up after the test."""
-        if not self._keep_data:
-            shutil.rmtree(settings.FLOW_EXECUTOR['DATA_DIR'], ignore_errors=True)
-
-        self.settings.disable()
-
-        super(TransactionTestCase, self).tearDown()
-
-    def keep_data(self, mock_purge=True):
-        """Do not delete output files after tests."""
-        self._keep_data = True
-
-        if mock_purge:
-            purge_mock_os = mock.patch('resolwe.flow.utils.purge.os', wraps=os).start()
-            purge_mock_os.remove = mock.MagicMock()
-
-            purge_mock_shutil = mock.patch('resolwe.flow.utils.purge.shutil', wraps=shutil).start()
-            purge_mock_shutil.rmtree = mock.MagicMock()
-
-
-class TestCase(TransactionTestCase, DjangoTestCase):
-    """Base class for writing Resolwe tests.
-
-    It is based on :class:`~resolwe.test.TransactionTestCase` and
-    Django's :class:`~django.test.TestCase`.
-    The latter encloses the test code in a database transaction that is
-    rolled back at the end of the test.
-
-    """
-
-    pass
 
 
 @override_settings(FLOW_EXECUTOR=FLOW_EXECUTOR_SETTINGS)
@@ -708,279 +547,3 @@ class ProcessTestCase(TestCase):
             msg += "\n".join(data.process_warning)
 
         return msg
-
-
-@override_settings(CELERY_ALWAYS_EAGER=True)
-class ResolweAPITestCase(APITestCase):
-    """Base class for testing Resolwe REST API.
-
-    This class is derived from Django REST Framework's
-    :drf:`APITestCase <testing/#test-cases>` class and has implemented
-    some basic features that makes testing Resolwe API easier. These
-    features includes following functions:
-
-    .. automethod:: _get_list
-    .. automethod:: _get_detail
-    .. automethod:: _post
-    .. automethod:: _patch
-    .. automethod:: _delete
-    .. automethod:: _detail_permissions
-
-    It also has included 2 views made from referenced DRF's ``ViewSet``.
-    First mimic list view and has following links between request's
-    methods and ViewSet's methods:
-
-      *  ``GET`` -> ``list``
-      *  ``POST`` -> ``create``
-
-    Second mimic detail view and has following links between request's
-    methods and ViewSet's methods:
-
-      *  ``GET`` -> ``retrieve``
-      *  ``PUT`` -> ``update``
-      *  ``PATCH`` -> ``partial_update``
-      *  ``DELETE`` -> ``destroy``
-      *  ``POST`` -> ``permissions``
-
-    If any of the listed methods is not defined in the VievSet,
-    corresponding link is omitted.
-
-    .. note::
-        ``self.viewset`` (instance of DRF's ``Viewset``) and
-        ``self.resource_name`` (string) must be defined before calling
-        super ``setUp`` method to work properly.
-
-    ``self.factory`` is instance of DRF's ``APIRequestFactory``.
-
-    """
-
-    def setUp(self):
-        """Prepare data."""
-        super(ResolweAPITestCase, self).setUp()
-
-        # TODO: Remove this when removing fixtures
-        if get_user_model().objects.filter(pk=2).exists():
-            self.user1 = get_user_model().objects.get(pk=2)
-        if get_user_model().objects.filter(pk=3).exists():
-            self.user2 = get_user_model().objects.get(pk=3)
-        if get_user_model().objects.filter(pk=4).exists():
-            self.user3 = get_user_model().objects.get(pk=4)
-        if get_user_model().objects.filter(pk=5).exists():
-            self.admin = get_user_model().objects.get(pk=5)
-
-        user_model = get_user_model()
-        # TODO: Change username to `admin` when fixtures are removed
-        self.admin = user_model.objects.create_superuser(username='admin2', email='admin@test.com', password='admin')
-        self.contributor = user_model.objects.create_user(username='contributor')
-
-        if not hasattr(self, 'viewset'):
-            raise KeyError("`self.viewset` must be defined in child class")
-
-        if not hasattr(self, 'resource_name'):
-            raise KeyError("`self.resource_name` must be defined in child class")
-
-        self.factory = APIRequestFactory()
-
-        list_url_mapping = {}
-        if hasattr(self.viewset, 'list'):  # pylint: disable=no-member
-            list_url_mapping['get'] = 'list'
-        if hasattr(self.viewset, 'create'):  # pylint: disable=no-member
-            list_url_mapping['post'] = 'create'
-
-        self.list_view = self.viewset.as_view(list_url_mapping)  # pylint: disable=no-member
-
-        detail_url_mapping = {}
-        if hasattr(self.viewset, 'retrieve'):  # pylint: disable=no-member
-            detail_url_mapping['get'] = 'retrieve'
-        if hasattr(self.viewset, 'update'):  # pylint: disable=no-member
-            detail_url_mapping['put'] = 'update'
-        if hasattr(self.viewset, 'partial_update'):  # pylint: disable=no-member
-            detail_url_mapping['patch'] = 'partial_update'
-        if hasattr(self.viewset, 'destroy'):  # pylint: disable=no-member
-            detail_url_mapping['delete'] = 'destroy'
-        if hasattr(self.viewset, 'detail_permissions'):  # pylint: disable=no-member
-            detail_url_mapping['post'] = 'detail_permissions'
-
-        self.detail_view = self.viewset.as_view(detail_url_mapping)  # pylint: disable=no-member
-
-    def detail_url(self, pk):
-        """Get detail url."""
-        return reverse('resolwe-api:{}-detail'.format(self.resource_name), kwargs={'pk': pk})  # noqa pylint: disable=no-member
-
-    def detail_permissions(self, pk):
-        """Get detail permissions url."""
-        return reverse('resolwe-api:{}-permissions'.format(self.resource_name), kwargs={'pk': pk})  # noqa pylint: disable=no-member
-
-    @property
-    def list_url(self):
-        """Get list url."""
-        return reverse('resolwe-api:{}-list'.format(self.resource_name))  # pylint: disable=no-member
-
-    def _render_query_params(self, params):
-        """Generate query parameters from given dict."""
-        if not params:
-            return ''
-
-        return '?' + '&'.join('{}={}'.format(key, value) for key, value in six.iteritems(params))
-
-    def _get_list(self, user=None, query_params={}):
-        """Make ``GET`` request to ``self.list_view`` view.
-
-        If ``user`` is not ``None``, the given user is authenticated
-        before making the request.
-
-        :param user: User to authenticate in request
-        :type user: :class:`~django.contrib.auth.models.User` or :data:`None`
-        :return: API response object
-        :rtype: :drf:`Response <responses/#response>`
-
-        """
-        url = self.list_url + self._render_query_params(query_params)
-        request = self.factory.get(url, format='json')
-        force_authenticate(request, user)
-        return self.list_view(request)
-
-    def _get_detail(self, pk, user=None, query_params={}):
-        """Make ``GET`` request to ``self.detail_view`` view.
-
-        If ``user`` is not ``None``, the given user is authenticated
-        before making the request.
-
-        :param int pk: Primary key of the coresponding object
-        :param user: User to authenticate in request
-        :type user: :class:`~django.contrib.auth.models.User` or :data:`None`
-        :return: API response object
-        :rtype: :drf:`Response <responses/#response>`
-
-        """
-        url = self.detail_url(pk) + self._render_query_params(query_params)
-        request = self.factory.get(url, format='json')
-        force_authenticate(request, user)
-        return self.detail_view(request, pk=pk)
-
-    def _post(self, data={}, user=None, query_params={}):
-        """Make ``POST`` request to ``self.list_view`` view.
-
-        If ``user`` is not ``None``, the given user is authenticated
-        before making the request.
-
-        :param dict data: data for posting in request's body
-        :param user: User to authenticate in request
-        :type user: :class:`~django.contrib.auth.models.User` or :data:`None`
-        :return: API response object
-        :rtype: :drf:`Response <responses/#response>`
-
-        """
-        url = self.list_url + self._render_query_params(query_params)
-        request = self.factory.post(url, data=data, format='json')
-        force_authenticate(request, user)
-        return self.list_view(request)
-
-    def _patch(self, pk, data={}, user=None, query_params={}):
-        """Make ``PATCH`` request to ``self.detail_view`` view.
-
-        If ``user`` is not ``None``, the given user is authenticated
-        before making the request.
-
-        :param int pk: Primary key of the coresponding object
-        :param dict data: data for posting in request's body
-        :param user: User to authenticate in request
-        :type user: :class:`~django.contrib.auth.models.User` or :data:`None`
-        :return: API response object
-        :rtype: :drf:`Response <responses/#response>`
-
-        """
-        url = self.detail_url(pk) + self._render_query_params(query_params)
-        request = self.factory.patch(url, data=data, format='json')
-        force_authenticate(request, user)
-        return self.detail_view(request, pk=pk)
-
-    def _delete(self, pk, user=None, query_params={}):
-        """Make ``DELETE`` request to ``self.detail_view`` view.
-
-        If ``user`` is not ``None``, the given user is authenticated
-        before making the request.
-
-        :param int pk: Primary key of the coresponding object
-        :param user: User to authenticate in request
-        :type user: :class:`~django.contrib.auth.models.User` or :data:`None`
-        :return: API response object
-        :rtype: :drf:`Response <responses/#response>`
-
-        """
-        url = self.detail_url(pk) + self._render_query_params(query_params)
-        request = self.factory.delete(url, format='json')
-        force_authenticate(request, user)
-        return self.detail_view(request, pk=pk)
-
-    def _detail_permissions(self, pk, data={}, user=None):
-        """Make ``POST`` request to ``self.detail_view`` view.
-
-        If ``user`` is not ``None``, the given user is authenticated
-        before making the request.
-
-        :param int pk: Primary key of the coresponding object
-        :param dict data: data for posting in request's body
-        :param user: User to authenticate in request
-        :type user: :class:`~django.contrib.auth.models.User` or :data:`None`
-        :return: API response object
-        :rtype: :drf:`Response <responses/#response>`
-
-        """
-        request = self.factory.post(self.detail_permissions(pk), data=data, format='json')
-        force_authenticate(request, user)
-        return self.detail_view(request, pk=pk)
-
-    def assertKeys(self, data, wanted):  # pylint: disable=invalid-name
-        """Assert dictionary keys."""
-        self.assertEqual(sorted(data.keys()), sorted(wanted))
-
-
-@override_settings(ELASTICSEARCH_INDEX_PREFIX='test')
-class TransactionElasticSearchTestCase(DjangoTransactionTestCase):
-    """Class for writing ElasticSearch tests not enclosed in a transaction.
-
-    It is based on Django's :class:`~django.test.TransactionTestCase`.
-    Use it if your tests depend on ElasticSearch and you need to access
-    the test's database from another thread/process.
-    This test class also takes care of cleaning the ElasticSearch data
-    before and after each test and prepares a fresh index.
-
-    """
-
-    def setUp(self):
-        """Delete any existing data and prepare fresh indexes."""
-        from resolwe.elastic.builder import index_builder, IndexBuilder
-        # Connect all signals
-        from resolwe.elastic import signals
-
-        super(TransactionElasticSearchTestCase, self).setUp()
-
-        index_builder.destroy()  # Delete default indexes
-
-        self._test_index_builder = IndexBuilder()
-        # Mock index builder
-        signals.index_builder = self._test_index_builder
-
-    def tearDown(self):
-        """Delete indexes and data from ElasticSearch."""
-        self._test_index_builder.destroy()
-        super(TransactionElasticSearchTestCase, self).tearDown()
-
-    def push_indexes(self):
-        """Push documents to Elasticsearch."""
-        self._test_index_builder.push(index=None)
-
-
-@override_settings(ELASTICSEARCH_INDEX_PREFIX='test')
-class ElasticSearchTestCase(DjangoTestCase, TransactionElasticSearchTestCase):
-    """Class for writing ElasticSearch tests.
-
-    It is based on :class:`~resolwe.test.TransactionElasticSearchTestCase`
-    and Django's :class:`~django.test.TestCase`.
-    The latter encloses the test code in a database transaction that is
-    rolled back at the end of the test.
-
-    """
-
-    pass
